requirements.txt
bidpom
PyPOM
pytest==2.9.2
pytest-selenium
pytest-xdist==1.14
